requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiohttp-validate==1.1.0
    # via -r requirements.in
aiohttp==3.7.3
    # via
    #   -r requirements.in
    #   aiohttp-validate
    #   pytest-aiohttp
aiopg==1.0.0
    # via -r requirements.in
aioredis==1.3.1
    # via -r requirements.in
alembic==1.4.3
    # via -r requirements.in
async-timeout==3.0.1
    # via
    #   aiohttp
    #   aioredis
attrs==20.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   pytest
certifi==2020.12.5
    # via requests
chardet==3.0.4
    # via
    #   aiohttp
    #   requests
docker==4.4.0
    # via -r requirements.in
hiredis==1.1.0
    # via aioredis
idna==2.10
    # via
    #   requests
    #   yarl
iniconfig==1.1.1
    # via pytest
jsonschema==3.2.0
    # via aiohttp-validate
mako==1.1.3
    # via alembic
markupsafe==1.1.1
    # via mako
multidict==5.0.2
    # via
    #   aiohttp
    #   yarl
packaging==20.7
    # via pytest
passlib==1.7.4
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
psycopg2-binary==2.8.6
    # via aiopg
psycopg2==2.8.6
    # via -r requirements.in
py==1.9.0
    # via pytest
pyjwt==1.7.1
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
pyrsistent==0.17.3
    # via jsonschema
pytest-aiohttp==0.3.0
    # via -r requirements.in
pytest==6.1.2
    # via
    #   -r requirements.in
    #   pytest-aiohttp
python-dateutil==2.8.1
    # via alembic
python-editor==1.0.4
    # via alembic
pyyaml==5.3.1
    # via trafaret-config
redis==3.5.3
    # via -r requirements.in
requests==2.25.0
    # via docker
six==1.15.0
    # via
    #   docker
    #   jsonschema
    #   python-dateutil
    #   sqlalchemy-utils
    #   websocket-client
sqlalchemy-utils==0.36.8
    # via -r requirements.in
sqlalchemy==1.3.20
    # via
    #   -r requirements.in
    #   alembic
    #   sqlalchemy-utils
toml==0.10.2
    # via pytest
trafaret-config==2.0.2
    # via -r requirements.in
trafaret==2.1.0
    # via
    #   -r requirements.in
    #   trafaret-config
typing-extensions==3.7.4.3
    # via aiohttp
urllib3==1.26.2
    # via requests
websocket-client==0.57.0
    # via docker
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
